LCSolution/LCSolutionTests/LongestPalindromeTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Solution;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Solution.Tests
{
    [TestClass()]
    public class LongestPalindromeTests
    {
        [TestMethod()]
        public void IsPalindromeStringTest()
        {
            LongestPalindrome sl = new LongestPalindrome();

            string input = string.Empty;
            bool expected = false;
            bool actual = false;

            input = "abbaass";
            expected = false;
            actual = sl.IsPalindromeString(input);
            Assert.AreEqual(expected, actual);

            input = "abba";
            expected = true;
            actual = sl.IsPalindromeString(input);
            Assert.AreEqual(expected, actual);
        }

        [TestMethod()]
        public void LongestPalindromeWrapperTest()
        {
            LongestPalindrome sl = new LongestPalindrome();
            string input = string.Empty;
            string expected = string.Empty;
            string actual = string.Empty;

            //input = "abba";
            //expected = "abba";
            //actual = sl.LongestPalindromeWrapper(input);
            //Assert.AreEqual(expected, actual);

            //input = "asabbaggb";
            //expected = "abba";
            //actual = sl.LongestPalindromeWrapper(input);
            //Assert.AreEqual(expected, actual);

            //input = "deamanaplanacanalpanama!amo";
            //expected = "amanaplanacanalpanama";
            //actual = sl.LongestPalindromeWrapper(input);
            //Assert.AreEqual(expected, actual);

            //input = "esbtzjaaijqkgmtaajpsdfiqtvxsgfvijpxrvxgfumsuprzlyvhclgkhccmcnquukivlpnjlfteljvykbddtrpmxzcrdqinsnlsteonhcegtkoszzonkwjevlasgjlcquzuhdmmkhfniozhuphcfkeobturbuoefhmtgcvhlsezvkpgfebbdbhiuwdcftenihseorykdguoqotqyscwymtjejpdzqepjkadtftzwebxwyuqwyeegwxhroaaymusddwnjkvsvrwwsmolmidoybsotaqufhepinkkxicvzrgbgsarmizugbvtzfxghkhthzpuetufqvigmyhmlsgfaaqmmlblxbqxpluhaawqkdluwfirfngbhdkjjyfsxglsnakskcbsyafqpwmwmoxjwlhjduayqyzmpkmrjhbqyhongfdxmuwaqgjkcpatgbrqdllbzodnrifvhcfvgbixbwywanivsdjnbrgskyifgvksadvgzzzuogzcukskjxbohofdimkmyqypyuexypwnjlrfpbtkqyngvxjcwvngmilgwbpcsseoywetatfjijsbcekaixvqreelnlmdonknmxerjjhvmqiztsgjkijjtcyetuygqgsikxctvpxrqtuhxreidhwcklkkjayvqdzqqapgdqaapefzjfngdvjsiiivnkfimqkkucltgavwlakcfyhnpgmqxgfyjziliyqhugphhjtlllgtlcsibfdktzhcfuallqlonbsgyyvvyarvaxmchtyrtkgekkmhejwvsuumhcfcyncgeqtltfmhtlsfswaqpmwpjwgvksvazhwyrzwhyjjdbphhjcmurdcgtbvpkhbkpirhysrpcrntetacyfvgjivhaxgpqhbjahruuejdmaghoaquhiafjqaionbrjbjksxaezosxqmncejjptcksnoq";
            //expected = "yvvy";
            //actual = sl.LongestPalindromeWrapper(input);
            //Assert.AreEqual(expected, actual);

            input = "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggggg";
            expected = input;
            actual = sl.LongestPalindromeWrapper(input);
            Assert.AreEqual(expected, actual);
        }
    }
}